TST: move fixtures used by stringdtype tests to test_stringdtype.py
