--- numpy/_core/tests/string_testing.py
@@ -1,10 +1,0 @@
-import string
-import numpy as np
-
-def random_string_list():
-    chars = list(string.ascii_letters + string.digits)
-    chars = np.array(chars, dtype="U1")
-    ret = np.random.choice(chars, size=100 * 10, replace=True)
-    return ret.view("U100")
-
-
